many many fixes and added codes

+++ routes/api.story.go
@@ -20,64 +20,13 @@ import (
 )
 
 // StoryRouter has api routes for stories
-func StoryRouter(app fiber.Router, userService user.Service, storyService story.Service, commentService comment.Service) {
-	api := app.Group("/api")
+func StoryRouter(api fiber.Router, userService user.Service, storyService story.Service, commentService comment.Service) {
 	api.Get("/blocks/:storyId", provideStoryBlocks(storyService))
 	api.Post("/story", middleware.APIGuard, addStory(userService, storyService))
 	api.Post("/like/:storyId/:plusMinus", middleware.APIGuard, handleLikeCount(userService, storyService))
 	api.Patch("/story/:storyId", middleware.APIGuard, editStory(storyService))
 	api.Delete("/story/:storyId", middleware.APIGuard, removeStory(userService, storyService, commentService))
-	admin := app.Group("/admin")
-	admin.Post("/pick/:storyId", pickStory(userService, storyService))
-	admin.Post("/unpick/:storyId", unpickStory(userService, storyService))
-}
 
-func unpickStory(userService user.Service, storyService story.Service) fiber.Handler {
-	return func(c *fiber.Ctx) error {
-
-		storyID := c.Params("storyId")
-		storyOID, err := primitive.ObjectIDFromHex(storyID)
-		if err != nil {
-			return c.SendStatus(500)
-		}
-		userOID, err := primitive.ObjectIDFromHex(fmt.Sprintf("%v", c.Locals("userId")))
-		if err != nil {
-			return c.SendStatus(500)
-		}
-		currentUser, err := userService.FindUserByID(userOID)
-		if !currentUser.IsEditor {
-			return c.SendStatus(403)
-		}
-		err = storyService.UnpickStory(storyOID)
-		if err != nil {
-			return c.SendStatus(500)
-		}
-		return c.SendStatus(200)
-	}
-}
-
-func pickStory(userService user.Service, storyService story.Service) fiber.Handler {
-	return func(c *fiber.Ctx) error {
-
-		storyID := c.Params("storyId")
-		storyOID, err := primitive.ObjectIDFromHex(storyID)
-		if err != nil {
-			return c.SendStatus(500)
-		}
-		userOID, err := primitive.ObjectIDFromHex(fmt.Sprintf("%v", c.Locals("userId")))
-		if err != nil {
-			return c.SendStatus(500)
-		}
-		currentUser, err := userService.FindUserByID(userOID)
-		if !currentUser.IsEditor {
-			return c.SendStatus(403)
-		}
-		err = storyService.PickStory(storyOID)
-		if err != nil {
-			return c.SendStatus(500)
-		}
-		return c.SendStatus(200)
-	}
 }
 
 func removeStory(userService user.Service, storyService story.Service, commentService comment.Service) fiber.Handler {
@@ -135,13 +84,15 @@ func removeStory(userService user.Service, storyService story.Service, commentSe
 			}
 		}
 
-		sess := myaws.ConnectAws()
-		svc := s3.New(sess)
-		bucketName := config.Config("BUCKET_NAME")
-		_, err = svc.DeleteObjects(&s3.DeleteObjectsInput{Bucket: aws.String(bucketName), Delete: &s3.Delete{Objects: objects, Quiet: aws.Bool(true)}})
-		if err != nil {
-			fmt.Println(err)
-			return c.SendStatus(500)
+		if len(objects) != 0 {
+			sess := myaws.ConnectAws()
+			svc := s3.New(sess)
+			bucketName := config.Config("BUCKET_NAME")
+			_, err = svc.DeleteObjects(&s3.DeleteObjectsInput{Bucket: aws.String(bucketName), Delete: &s3.Delete{Objects: objects, Quiet: aws.Bool(true)}})
+			if err != nil {
+				fmt.Println(err)
+				return c.SendStatus(500)
+			}
 		}
 
 		err = storyService.RemoveStory(storyOID)
